added weight connection

+++ JavaSpring/src/main/java/be/kdg/integration3/easyrep/api/controller/SetControllerAPI.java
@@ -44,7 +44,7 @@ public class SetControllerAPI {
 
 //        setService.addSet(startLocalTime, endLocalTime, repCount);
         logger.debug("Inputting values");
-        setService.addSet(setNumber, setTime, repCount, weight);
+        setService.addSet(setNumber, setTime, repCount,(double) weight);
     }
 
 }
